Externalize forward declarations for Track object
...to reduce compile time dependencies
