+++ src/test/beatgridtest.cpp
@@ -1,7 +1,9 @@
 #include <gtest/gtest.h>
+
 #include <QtDebug>
 
 #include "track/beatgrid.h"
+#include "track/track.h"
 #include "util/memory.h"
 
 using namespace mixxx;
