+++ src/track/beatmap.h
@@ -10,9 +10,9 @@
 
 #include <QMutex>
 
-#include "track/track.h"
-#include "track/beats.h"
 #include "proto/beats.pb.h"
+#include "track/beats.h"
+#include "track/track_decl.h"
 
 #define BEAT_MAP_VERSION "BeatMap-1.0"
