+++ src/track/globaltrackcache.cpp
@@ -2,6 +2,7 @@
 
 #include <QCoreApplication>
 
+#include "track/track.h"
 #include "util/assert.h"
 #include "util/logger.h"
 #include "util/thread_affinity.h"
