+++ src/track/globaltrackcache.h
@@ -4,9 +4,9 @@
 #include <map>
 #include <unordered_map>
 
-#include "track/track.h"
+#include "track/track_decl.h"
 #include "track/trackref.h"
-
+#include "util/sandbox.h"
 
 // forward declaration(s)
 class GlobalTrackCache;
