+++ src/waveform/renderers/glwaveformrenderersimplesignal.cpp
@@ -1,6 +1,7 @@
 #include "waveform/renderers/glwaveformrenderersimplesignal.h"
 #if !defined(QT_NO_OPENGL) && !defined(QT_OPENGL_ES_2)
 
+#include "track/track.h"
 #include "util/math.h"
 #include "waveform/waveform.h"
 #include "waveform/waveformwidgetfactory.h"
